Change the implementation of the removePhotos method
Before the pin's photos property was simply set to nil which lead to problems (crashes) because the relationship between a pin and its photos has to be bilateral. The photos property however was just set to nil but the photos weren't actually deleted. In this implementation the photos are deleted with a batch delete request.