Modify map view delegate methods for custom annotation view
The custom annotation view's images were blurry when resizing them manually which actually isn't necessary when the image for the custom annotation view is provided in 3 different sizes. Thus, instead of creating UIImage objects, image literals are used and not resized manually. Also, code for the photo album view controller's map view delegate was moved into an appropriately named file as it was mistakenly in a file that indicated it was a data source function.